requirements.txt
flask==3.0.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
Pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
umap-learn==0.5.5
